+++ vck-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/oidvci/CredentialSchemeMapping.kt
@@ -1,11 +1,8 @@
 package at.asitplus.wallet.lib.oidvci
 
-import at.asitplus.openid.ClaimDescription
 import at.asitplus.openid.CredentialFormatEnum
 import at.asitplus.openid.CredentialFormatEnum.DC_SD_JWT
 import at.asitplus.openid.CredentialFormatEnum.JWT_VC
-import at.asitplus.openid.OpenId4VciClaimsPathPointer
-import at.asitplus.openid.OpenId4VciClaimsPathPointerSegmentString
 import at.asitplus.openid.OpenIdConstants.BINDING_METHOD_COSE_KEY
 import at.asitplus.openid.OpenIdConstants.BINDING_METHOD_JWK
 import at.asitplus.openid.OpenIdConstants.URN_TYPE_JWK_THUMBPRINT
@@ -66,11 +63,7 @@ fun CredentialScheme.toIsoMdocSupportedCredentialFormat(identifier: String): Pai
         scope = identifier,
         docType = isoDocType!!,
         supportedBindingMethods = setOf(BINDING_METHOD_JWK, BINDING_METHOD_COSE_KEY),
-        isoClaims = claimNames.map {
-            ClaimDescription(path = OpenId4VciClaimsPathPointer(isoNamespace!!) + it.split(".").map {
-                OpenId4VciClaimsPathPointerSegmentString(it)
-            })
-        }.toSet()
+        isoClaims = claimDescriptions
     )
 
 fun CredentialScheme.toPlainJwtSupportedCredentialFormat(identifier: String): Pair<String, SupportedCredentialFormat> =
@@ -80,9 +73,7 @@ fun CredentialScheme.toPlainJwtSupportedCredentialFormat(identifier: String): Pa
             types = setOf(VcDataModelConstants.VERIFIABLE_CREDENTIAL, vcType!!),
         ),
         supportedBindingMethods = setOf(BINDING_METHOD_JWK, URN_TYPE_JWK_THUMBPRINT),
-        vcJwtClaims = claimNames.map {
-            ClaimDescription(path = OpenId4VciClaimsPathPointer(it.split(".")))
-        }.toSet()
+        vcJwtClaims = claimDescriptions
     )
 
 fun CredentialScheme.toSdJwtSupportedCredentialFormat(identifier: String): Pair<String, SupportedCredentialFormat> =
@@ -90,10 +81,7 @@ fun CredentialScheme.toSdJwtSupportedCredentialFormat(identifier: String): Pair<
         scope = identifier,
         sdJwtVcType = sdJwtType!!,
         supportedBindingMethods = setOf(BINDING_METHOD_JWK, URN_TYPE_JWK_THUMBPRINT),
-        sdJwtClaims = claimDescriptions.takeUnless { it.isEmpty() }?.toSet()
-            ?: claimNames.map {
-                ClaimDescription(path = OpenId4VciClaimsPathPointer(it.split(".")))
-            }.toSet()
+        sdJwtClaims = claimDescriptions
     )
 
 class DefaultCredentialSchemeMapper : CredentialSchemeMapper {