+++ vck-openid/src/commonTest/kotlin/at/asitplus/wallet/lib/oidvci/SameScopeCredentialSchemeMapper.kt
@@ -1,10 +1,8 @@
 package at.asitplus.wallet.lib.oidvci
 
-import at.asitplus.openid.ClaimDescription
 import at.asitplus.openid.CredentialFormatEnum
 import at.asitplus.openid.CredentialFormatEnum.DC_SD_JWT
 import at.asitplus.openid.CredentialFormatEnum.JWT_VC
-import at.asitplus.openid.OpenId4VciClaimsPathPointer
 import at.asitplus.openid.OpenIdConstants.BINDING_METHOD_COSE_KEY
 import at.asitplus.openid.OpenIdConstants.BINDING_METHOD_JWK
 import at.asitplus.openid.OpenIdConstants.URN_TYPE_JWK_THUMBPRINT
@@ -49,9 +47,7 @@ class SameScopeCredentialSchemeMapper(
                     scope = scope,
                     docType = isoDocType!!,
                     supportedBindingMethods = setOf(BINDING_METHOD_JWK, BINDING_METHOD_COSE_KEY),
-                    isoClaims = claimNames.map {
-                        ClaimDescription(path = OpenId4VciClaimsPathPointer(listOf(isoNamespace!!) + it.split(".")))
-                    }.toSet()
+                    isoClaims = claimDescriptions
                 )
             }
         } else null
@@ -63,9 +59,7 @@ class SameScopeCredentialSchemeMapper(
                         types = setOf(VcDataModelConstants.VERIFIABLE_CREDENTIAL, vcType!!),
                     ),
                     supportedBindingMethods = setOf(BINDING_METHOD_JWK, URN_TYPE_JWK_THUMBPRINT),
-                    vcJwtClaims = claimNames.map {
-                        ClaimDescription(path = OpenId4VciClaimsPathPointer(it.split(".")))
-                    }.toSet()
+                    vcJwtClaims = claimDescriptions
                 )
             }
         } else null
@@ -75,9 +69,7 @@ class SameScopeCredentialSchemeMapper(
                     scope = scope,
                     sdJwtVcType = sdJwtType!!,
                     supportedBindingMethods = setOf(BINDING_METHOD_JWK, URN_TYPE_JWK_THUMBPRINT),
-                    sdJwtClaims = claimNames.map {
-                        ClaimDescription(path = OpenId4VciClaimsPathPointer(it.split(".")))
-                    }.toSet()
+                    sdJwtClaims = claimDescriptions
                 )
             }
         } else null